Update settings section
This commit updates the settings section:

- Server name is not required anymore. We default it to `https://feedland.social`.
- Added validation for the `feedland_blogroll_options`: since we're dealing with an
array, it's a little bit tricky to add validation to each option individually since
the `sanitize_callback` is for the whole option. Two distinct validations were added,
one for the username and other for the category field. Both check the FeedLand
server before saving the option. If any error occurs, a notice is displayed and we
default to an empty string for the category and `davewiner` for the username.